+++ Shem/Shem.csproj
@@ -39,12 +39,12 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Controllers\TorController.cs" />
-    <Compile Include="Protocols\TC\Commands\AUTHENTICATE.cs" />
-    <Compile Include="Protocols\TC\Commands\GETCONF.cs" />
-    <Compile Include="Protocols\TC\Commands\GETINFO.cs" />
-    <Compile Include="Protocols\TC\Commands\SIGNAL.cs" />
-    <Compile Include="Protocols\TC\TCCommand.cs" />
+    <Compile Include="TorController.cs" />
+    <Compile Include="Commands\AUTHENTICATE.cs" />
+    <Compile Include="Commands\GETCONF.cs" />
+    <Compile Include="Commands\GETINFO.cs" />
+    <Compile Include="Commands\SIGNAL.cs" />
+    <Compile Include="Commands\TCCommand.cs" />
     <Compile Include="Sockets\ControlSocket.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Sockets\SocketNotConnectedException.cs" />
